requirements.txt
pygame==2.0
box2d==2.3.10
pillow==8.0.1
svgelements==1.3.5
